requirements.txt
# Core ML
torch>=2.1.0
transformers>=4.40.0
peft>=0.10.0
trl>=0.8.0
accelerate>=0.28.0
bitsandbytes>=0.43.0
datasets>=2.18.0

# Evaluation
openai>=1.12.0
anthropic>=0.18.0
google-genai>=0.4.0

# Analysis & Visualization
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
pandas>=2.1.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
jsonlines>=4.0.0
wandb>=0.16.0

# Code evaluation
evalplus>=0.3.0
